+++ src/main/java/kr/co/mdi/dao/DeviceDaoPostgreImpl.java
@@ -43,16 +43,16 @@ public class DeviceDaoPostgreImpl implements DeviceDao {
 
 			while (rs.next()) {
 				DeviceDTO device = new DeviceDTO();
-				device.setIdDevice(rs.getLong("id_device"));
+				device.setIdDevice(rs.getInt("id_device"));
 				device.setTypeDevice(rs.getString("type_device"));
 				device.setNameDevice(rs.getString("name_device"));
 				device.setManfDevice(rs.getString("manf_device"));
 				device.setDeviceManfCode(rs.getString("device_manf_code"));
 				device.setCpuDevice(rs.getString("cpu_device"));
 				device.setLineupDevice(rs.getString("lineup_device"));
-				device.setReleaseDevice(rs.getLong("release_device"));
+				device.setReleaseDevice(rs.getInt("release_device"));
 				device.setWeightDevice(rs.getFloat("weight_device"));
-				device.setChoiceDevice(rs.getLong("choice_device"));
+				device.setChoiceDevice(rs.getInt("choice_device"));
 
 				deviceList.add(device);
 			}
@@ -66,26 +66,26 @@ public class DeviceDaoPostgreImpl implements DeviceDao {
 	}
 
 	@Override
-	public DeviceDTO selectDeviceById(Long deviceId) {
+	public DeviceDTO selectDeviceById(Integer deviceId) {
 		String sql = "SELECT * FROM mdl WHERE id_device = ?";
 		DeviceDTO device = null;
 
 		try (Connection conn = getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
-			pstmt.setLong(1, deviceId);
+			pstmt.setInt(1, deviceId);
 			try (ResultSet rs = pstmt.executeQuery()) {
 				if (rs.next()) {
 					device = new DeviceDTO();
-					device.setIdDevice(rs.getLong("id_device"));
+					device.setIdDevice(rs.getInt("id_device"));
 					device.setTypeDevice(rs.getString("type_device"));
 					device.setNameDevice(rs.getString("name_device"));
 					device.setManfDevice(rs.getString("manf_device"));
 					device.setDeviceManfCode(rs.getString("device_manf_code"));
 					device.setCpuDevice(rs.getString("cpu_device"));
 					device.setLineupDevice(rs.getString("lineup_device"));
-					device.setReleaseDevice(rs.getLong("release_device"));
+					device.setReleaseDevice(rs.getInt("release_device"));
 					device.setWeightDevice(rs.getFloat("weight_device"));
-					device.setChoiceDevice(rs.getLong("choice_device"));
+					device.setChoiceDevice(rs.getInt("choice_device"));
 
 				}
 			}
