+++ src/main/java/kr/co/mdi/dto/DeviceDTO.java
@@ -2,12 +2,12 @@ package kr.co.mdi.dto;
 
 public class DeviceDTO {
 
-	private Long idDevice; // mdl.id_device
+	private Integer idDevice; // mdl.id_device
 	private String nameDevice; // mdl.name_device
 	private String lineupDevice; // mdl.lineup_device
-	private Long releaseDevice; // mdl.release_device
+	private Integer releaseDevice; // mdl.release_device
 	private Float weightDevice; // mdl.weight_device
-	private Long choiceDevice; // mdl.choice_device
+	private Integer choiceDevice; // mdl.choice_device
 
 	private String deviceTypeCode; // mdl.device_type_code (외래 키)
 	private String typeDevice; // device_type.type_device (표시용 이름)
@@ -15,14 +15,14 @@ public class DeviceDTO {
 	private String deviceManfCode; // mdl.device_manf_code (외래 키)
 	private String manfDevice; // device_manf_brand.manf_device (표시용 이름)
 
-	private Long idCpu; // mdl.id_cpu (외래 키 → mcl.id_cpu)
+	private Integer idCpu; // mdl.id_cpu (외래 키 → mcl.id_cpu)
 	private String cpuDevice; // mcl.name_cpu (표시용 CPU 이름)
 
-	public Long getIdDevice() {
+	public Integer getIdDevice() {
 		return idDevice;
 	}
 
-	public void setIdDevice(Long idDevice) {
+	public void setIdDevice(Integer idDevice) {
 		this.idDevice = idDevice;
 	}
 
@@ -42,11 +42,11 @@ public class DeviceDTO {
 		this.lineupDevice = lineupDevice;
 	}
 
-	public Long getReleaseDevice() {
+	public Integer getReleaseDevice() {
 		return releaseDevice;
 	}
 
-	public void setReleaseDevice(Long releaseDevice) {
+	public void setReleaseDevice(Integer releaseDevice) {
 		this.releaseDevice = releaseDevice;
 	}
 
@@ -58,11 +58,11 @@ public class DeviceDTO {
 		this.weightDevice = weightDevice;
 	}
 
-	public Long getChoiceDevice() {
+	public Integer getChoiceDevice() {
 		return choiceDevice;
 	}
 
-	public void setChoiceDevice(Long choiceDevice) {
+	public void setChoiceDevice(Integer choiceDevice) {
 		this.choiceDevice = choiceDevice;
 	}
 
@@ -98,11 +98,11 @@ public class DeviceDTO {
 		this.manfDevice = manfDevice;
 	}
 
-	public Long getIdCpu() {
+	public Integer getIdCpu() {
 		return idCpu;
 	}
 
-	public void setIdCpu(Long idCpu) {
+	public void setIdCpu(Integer idCpu) {
 		this.idCpu = idCpu;
 	}
