+++ src/main/java/kr/co/mdi/service/CpuService.java
@@ -7,6 +7,6 @@ import kr.co.mdi.dto.CpuDTO;
 public interface CpuService {
 	List<CpuDTO> getCpuList();
 
-	CpuDTO getCpuById(Long cpuId);
+	CpuDTO getCpuById(Integer cpuId);
 
 }
